Flag to disable HasLens instances for tuples
