--- src/Ether/Internal/TupleInstances_HasLens_4.hs
@@ -1,7 +1,0 @@
-{-# OPTIONS -fno-warn-orphans #-}
-
-module Ether.Internal.TupleInstances_HasLens_4 () where
-
-import Ether.Internal.TH_TupleInstances (makeTupleInstancesHasLens)
-
-makeTupleInstancesHasLens [37..42]
