--- src/Ether/Internal/TupleInstances_HasLens_8.hs
@@ -1,7 +1,0 @@
-{-# OPTIONS -fno-warn-orphans #-}
-
-module Ether.Internal.TupleInstances_HasLens_8 () where
-
-import Ether.Internal.TH_TupleInstances (makeTupleInstancesHasLens)
-
-makeTupleInstancesHasLens [57..60]
